refactor(#21): added/split common package
- utils
- exception
