+++ archive-common/src/main/java/site/archive/common/DateTimeUtil.java
@@ -1,6 +1,4 @@
-package site.archive.util;
-
-import lombok.extern.slf4j.Slf4j;
+package site.archive.common;
 
 import java.time.Clock;
 import java.time.Instant;
@@ -11,7 +9,6 @@ import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 
-@Slf4j
 public class DateTimeUtil {
 
     public static final DateTimeFormatter YY_MM_DD_FORMATTER = DateTimeFormatter.ofPattern("yy/MM/dd");
@@ -24,7 +21,7 @@ public class DateTimeUtil {
     private DateTimeUtil() {}
 
     public static void changeClock(LocalDate date) {
-        log.warn("Change clock of DateTimeUtil. It may effect now() methods of DateTimeUtil.");
+        // log.warn("Change clock of DateTimeUtil. It may effect now() methods of DateTimeUtil.");
         var dateTime = date.atStartOfDay();
         var zoneOffset = ASIA_SEOUL_ZONE.getRules().getOffset(dateTime);
         clock = Clock.fixed(dateTime.atOffset(zoneOffset).toInstant(), zoneOffset);
