+++ archive-common/src/main/java/site/archive/common/exception/common/UnauthorizedResourceException.java
@@ -1,7 +1,7 @@
-package site.archive.exception.common;
+package site.archive.common.exception.common;
 
-import site.archive.exception.BaseException;
-import site.archive.exception.ExceptionCode;
+import site.archive.common.exception.BaseException;
+import site.archive.common.exception.ExceptionCode;
 
 public class UnauthorizedResourceException extends BaseException {
