+++ archive-common/src/main/java/site/archive/common/exception/security/InvalidTokenException.java
@@ -1,7 +1,7 @@
-package site.archive.exception.security;
+package site.archive.common.exception.security;
 
-import site.archive.exception.BaseException;
-import site.archive.exception.ExceptionCode;
+import site.archive.common.exception.BaseException;
+import site.archive.common.exception.ExceptionCode;
 
 public class InvalidTokenException extends BaseException {
