+++ archive-common/src/main/java/site/archive/common/exception/security/TokenNotFoundException.java
@@ -1,0 +1,11 @@
+package site.archive.common.exception.security;
+
+import site.archive.common.exception.ExceptionCode;
+
+public class TokenNotFoundException extends RuntimeException {
+
+    public TokenNotFoundException() {
+        super(ExceptionCode.TOKEN_NOT_FOUND.getMessage());
+    }
+
+}
