+++ archive-common/src/main/java/site/archive/common/exception/security/WrappingAuthenticationException.java
@@ -1,0 +1,16 @@
+package site.archive.common.exception.security;
+
+import lombok.Getter;
+import site.archive.common.exception.BaseException;
+
+@Getter
+public class WrappingAuthenticationException extends RuntimeException {
+
+    private final BaseException cause;
+
+    public WrappingAuthenticationException(BaseException cause) {
+        super(cause.getMessage());
+        this.cause = cause;
+    }
+
+}
