+++ archive-common/src/test/java/site/archive/common/DateTimeUtilTest.java
@@ -1,4 +1,4 @@
-package site.archive.util;
+package site.archive.common;
 
 import org.junit.jupiter.api.Test;
