+++ settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'archive'
+
+include 'archive-domain'
+include 'archive-common'
