+++ src/main/java/site/archive/api/v1/UserControllerV1.java
@@ -16,10 +16,10 @@ import site.archive.api.resolver.annotation.RequestUser;
 import site.archive.api.v1.dto.archive.EmailDuplicateResponseDto;
 import site.archive.api.v1.dto.user.UserEmailRequestDto;
 import site.archive.api.v1.dto.user.UserPasswordResetRequestDto;
+import site.archive.config.security.util.SecurityUtils;
 import site.archive.domain.user.UserAuthService;
 import site.archive.domain.user.UserService;
 import site.archive.domain.user.info.UserInfo;
-import site.archive.util.SecurityUtils;
 
 @RestController
 @RequestMapping("/api/v1/auth")
