+++ src/main/java/site/archive/api/v1/dto/user/BaseUserDto.java
@@ -2,8 +2,8 @@ package site.archive.api.v1.dto.user;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import site.archive.common.DateTimeUtil;
 import site.archive.domain.user.entity.BaseUser;
-import site.archive.util.DateTimeUtil;
 
 import java.time.LocalDateTime;
