+++ src/main/java/site/archive/config/security/util/SecurityUtils.java
@@ -1,9 +1,9 @@
-package site.archive.util;
+package site.archive.config.security.util;
 
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.context.SecurityContextHolder;
-import site.archive.domain.user.info.UserInfo;
 import site.archive.config.security.token.jwt.JwtAuthenticationToken;
+import site.archive.domain.user.info.UserInfo;
 
 import java.security.SecureRandom;
