+++ src/main/java/site/archive/domain/archive/ArchiveCommunityTimeSortType.java
@@ -3,9 +3,9 @@ package site.archive.domain.archive;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import lombok.Getter;
+import site.archive.common.DateTimeUtil;
 import site.archive.domain.archive.entity.Archive;
 import site.archive.domain.archive.entity.QArchive;
-import site.archive.util.DateTimeUtil;
 
 import java.util.Arrays;
 import java.util.HashMap;
