+++ src/main/java/site/archive/domain/archive/ArchiveCustomRepositoryImpl.java
@@ -5,9 +5,9 @@ import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
+import site.archive.common.DateTimeUtil;
 import site.archive.domain.archive.entity.Archive;
 import site.archive.domain.archive.entity.Emotion;
-import site.archive.util.DateTimeUtil;
 
 import java.util.List;
