+++ src/main/java/site/archive/domain/user/UserRegisterService.java
@@ -8,10 +8,10 @@ import org.springframework.transaction.annotation.Transactional;
 import site.archive.api.command.BasicRegisterCommand;
 import site.archive.api.command.OAuthRegisterCommand;
 import site.archive.api.v1.dto.user.BaseUserDto;
+import site.archive.common.exception.common.DuplicateResourceException;
 import site.archive.domain.common.MessagingService;
 import site.archive.domain.user.entity.BaseUser;
 import site.archive.domain.user.info.UserInfo;
-import site.archive.exception.common.DuplicateResourceException;
 
 @Service
 @Transactional
