+++ src/main/java/site/archive/infra/cloud/aws/S3Service.java
@@ -6,8 +6,8 @@ import com.amazonaws.services.s3.model.ObjectMetadata;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import site.archive.common.exception.infra.FileInvalidException;
 import site.archive.domain.archive.ArchiveImageService;
-import site.archive.exception.infra.FileInvalidException;
 import site.archive.infra.cloud.aws.config.AwsS3Property;
 
 import java.io.IOException;
